# api/openapi.yaml
openapi: 3.1.0
info:
  title: KeebHub API
  version: 0.1.0
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
  description: >-
    Public HTTP contract for KeebHub v1. Health, readiness, Discord authentication,
    logout, current-user profile, seller catalog, and catalog foundation are implemented. Operations marked
    planned are implemented by later vertical slices without changing this contract unless a
    reviewed contract change is made.
servers:
  - url: /
...
security: []
paths:
...
  /auth/discord:
    get:
      x-implementation-status: implemented
      tags: [Authentication]
      operationId: startDiscordLogin
      summary: Start Discord OAuth2 login
      parameters:
        - name: return_to
          in: query
          required: false
          description: Relative internal application path.
          schema:
            type: string
            default: /
            maxLength: 2048
      responses:
        "302":
          description: Redirect to Discord authorization.
          headers:
            Location:
              required: true
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/discord/callback:
    get:
      x-implementation-status: implemented
      tags: [Authentication]
      operationId: finishDiscordLogin
      summary: Finish Discord OAuth2 login
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
            minLength: 1
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: error
          in: query
          required: false
          description: Discord authorization error, including user denial.
          schema:
            type: string
            minLength: 1
      responses:
        "302":
          description: Session created and redirected internally, or failure redirected to the login page.
          headers:
            Location:
              required: true
              schema:
                type: string
  /auth/logout:
    post:
      x-implementation-status: implemented
      tags: [Authentication]
      operationId: logout
      summary: Invalidate the current session
      responses:
        "204":
          description: Session invalidated or already absent.
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/me:
    get:
      x-implementation-status: implemented
      tags: [Profile]
      operationId: getCurrentUser
      summary: Get the authenticated user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      x-implementation-status: implemented
      tags: [Profile]
      operationId: updateCurrentUser
      summary: Update optional profile fields
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated current user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
...
  /api/v1/listings:
    get:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: searchListings
      summary: Search public active and reserved listings
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/ConditionFilter"
        - $ref: "#/components/parameters/MinimumPrice"
        - $ref: "#/components/parameters/MaximumPrice"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Matching public listings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: createListing
      summary: Create a listing
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Listing created.
          content:
            application/json:
              schema:
                type: object
                required: [listing]
                properties:
                  listing:
                    $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /api/v1/listings/{listing_id}:
    parameters:
      - $ref: "#/components/parameters/ListingID"
    get:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: getListing
      summary: Get a listing subject to visibility rules
      responses:
        "200":
          description: Listing details.
          content:
            application/json:
              schema:
                type: object
                required: [listing]
                properties:
                  listing:
                    $ref: "#/components/schemas/Listing"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: updateListing
      summary: Update an owned listing
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Listing updated.
          content:
            application/json:
              schema:
                type: object
                required: [listing]
                properties:
                  listing:
                    $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /api/v1/listings/{listing_id}/status:
    parameters:
      - $ref: "#/components/parameters/ListingID"
    post:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: changeListingStatus
      summary: Apply an allowed listing status transition
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/ListingStatus"
      responses:
        "200":
          description: Listing status changed.
          content:
            application/json:
              schema:
                type: object
                required: [listing]
                properties:
                  listing:
                    $ref: "#/components/schemas/Listing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/users/{handle}:
    get:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: getSellerProfile
      summary: Get a public seller profile
      parameters:
        - $ref: "#/components/parameters/Handle"
      responses:
        "200":
          description: Public seller profile.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/SellerProfile"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{handle}/listings:
    get:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: listSellerListings
      summary: List public seller listings
      parameters:
        - $ref: "#/components/parameters/Handle"
        - $ref: "#/components/parameters/PublicStatusFilter"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Seller listings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/listings:
    get:
      x-implementation-status: implemented
      tags: [Catalog]
      operationId: listOwnedListings
      summary: List the current user's listings
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/OwnedStatusFilter"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Owned listings including seller-controlled states.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/listings/{listing_id}/conversation:
    parameters:
      - $ref: "#/components/parameters/ListingID"
    post:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: startConversation
      summary: Create or return the buyer's listing conversation
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Existing conversation returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationCreated"
        "201":
          description: Conversation created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationCreated"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/conversations:
    get:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: listConversations
      summary: List the current user's inbox
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Inbox page.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConversationSummary"
                  next_cursor:
                    $ref: "#/components/schemas/NullableCursor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/conversations/{conversation_id}/messages:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
    get:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: listMessages
      summary: Read or catch up conversation messages
      security:
        - sessionCookie: []
      parameters:
        - name: before_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ID"
        - name: after_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ID"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Conversation messages.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: sendMessage
      summary: Persist a conversation message
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        "201":
          description: Message committed before realtime publication.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/conversations/{conversation_id}/read:
    parameters:
      - $ref: "#/components/parameters/ConversationID"
    post:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: markConversationRead
      summary: Advance the current user's read pointer
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [last_read_message_id]
              properties:
                last_read_message_id:
                  $ref: "#/components/schemas/ID"
      responses:
        "204":
          description: Read pointer advanced or already ahead.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /api/v1/events:
    get:
      x-implementation-status: implemented
      tags: [Chat]
      operationId: streamEvents
      summary: Stream best-effort user events
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Server-Sent Events stream with message.created events.
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: keebhub_session
  parameters:
    ListingID:
      name: listing_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    ConversationID:
      name: conversation_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"
    Handle:
      name: handle
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$"
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
        maxLength: 100
    CategoryFilter:
      name: category
      in: query
      required: false
      schema:
        type: string
        maxLength: 50
    ConditionFilter:
      name: condition
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ListingCondition"
    MinimumPrice:
      name: min_price
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000000000
    MaximumPrice:
      name: max_price
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000000000
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [newest, price_asc, price_desc]
        default: newest
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to the filters and sort that created it.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PublicStatusFilter:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [active, reserved, sold]
    OwnedStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/ListingStatus"
  responses:
    BadRequest:
      description: Malformed request or invalid cursor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: A valid active session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated caller may not perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Current resource state prevents the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: Domain validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RequestTooLarge:
      description: The request body exceeds the endpoint limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnsupportedMediaType:
      description: The request body must use application/json.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Request rate exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required local dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ID:
      type: string
      pattern: "^[1-9][0-9]*$"
      examples: ["1001"]
    NullableCursor:
      anyOf:
        - type: string
          minLength: 1
        - type: "null"
    Timestamp:
      type: string
      format: date-time
      examples: ["2026-08-08T12:00:00Z"]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              pattern: "^[a-z][a-z0-9_]*$"
            message:
              type: string
              minLength: 1
            request_id:
              type: string
              minLength: 1
            fields:
              type: object
              additionalProperties:
                type: string
    User:
      type: object
      additionalProperties: false
      required: [id, handle, discord_username, display_name, created_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        handle:
          type: string
        discord_username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: [string, "null"]
          format: uri
        location:
          type: [string, "null"]
          maxLength: 100
        bio:
          type: [string, "null"]
          maxLength: 500
        created_at:
          $ref: "#/components/schemas/Timestamp"
    PublicUser:
      type: object
      additionalProperties: false
      required: [id, handle, display_name]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        handle:
          type: string
        display_name:
          type: string
        avatar_url:
          type: [string, "null"]
          format: uri
        location:
          type: [string, "null"]
          maxLength: 100
        bio:
          type: [string, "null"]
          maxLength: 500
    SellerProfile:
      type: object
      additionalProperties: false
      required: [id, handle, display_name, avatar_url, location, bio, created_at, active_listing_count]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        handle:
          type: string
        display_name:
          type: string
        avatar_url:
          type: [string, "null"]
          format: uri
        location:
          type: [string, "null"]
          maxLength: 100
        bio:
          type: [string, "null"]
          maxLength: 500
        created_at:
          $ref: "#/components/schemas/Timestamp"
        active_listing_count:
          type: integer
          minimum: 0
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        location:
          type: [string, "null"]
          maxLength: 100
        bio:
          type: [string, "null"]
          maxLength: 500
    Category:
      type: object
      additionalProperties: false
      required: [id, slug, name]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        slug:
          type: string
        name:
          type: string
    ListingCondition:
      type: string
      enum: [new, used]
    ListingStatus:
      type: string
      enum: [active, reserved, sold, archived]
    Listing:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - description
        - price_idr
        - quantity
        - category
        - condition
        - status
        - negotiable
        - seller
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/ID"
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 5000
        price_idr:
          type: integer
          minimum: 1
          maximum: 10000000000
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        category:
          $ref: "#/components/schemas/Category"
        condition:
          $ref: "#/components/schemas/ListingCondition"
        status:
          $ref: "#/components/schemas/ListingStatus"
        negotiable:
          type: boolean
        seller:
          $ref: "#/components/schemas/PublicUser"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    ListingPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        next_cursor:
          $ref: "#/components/schemas/NullableCursor"
    CreateListingRequest:
      type: object
      additionalProperties: false
      required: [title, price_idr, category_slug, condition]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 5000
          default: ""
        price_idr:
          type: integer
          minimum: 1
          maximum: 10000000000
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
          default: 1
        category_slug:
          type: string
          minLength: 1
          maxLength: 50
        condition:
          $ref: "#/components/schemas/ListingCondition"
        negotiable:
          type: boolean
          default: false
    UpdateListingRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 5000
        price_idr:
          type: integer
          minimum: 1
          maximum: 10000000000
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        category_slug:
          type: string
          minLength: 1
          maxLength: 50
        condition:
          $ref: "#/components/schemas/ListingCondition"
        negotiable:
          type: boolean
    ConversationCreated:
      type: object
      additionalProperties: false
      required: [conversation, created]
      properties:
        conversation:
          type: object
          additionalProperties: false
          required: [id]
          properties:
            id:
              $ref: "#/components/schemas/ID"
        created:
          type: boolean
    ConversationSummary:
      type: object
      additionalProperties: false
      required: [id, listing, counterpart, unread_count]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        listing:
          type: object
          additionalProperties: false
          required: [id, title, status]
          properties:
            id:
              $ref: "#/components/schemas/ID"
            title:
              type: string
            status:
              $ref: "#/components/schemas/ListingStatus"
        counterpart:
          $ref: "#/components/schemas/PublicUser"
        last_message:
          anyOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        unread_count:
          type: integer
          minimum: 0
    Message:
      type: object
      additionalProperties: false
      required: [id, sender_id, body, created_at]
      properties:
        id:
          $ref: "#/components/schemas/ID"
        sender_id:
          $ref: "#/components/schemas/ID"
        body:
          type: string
          minLength: 1
          maxLength: 2000
        created_at:
          $ref: "#/components/schemas/Timestamp"